+++ scripts/addBranchConfig.js
@@ -2,9 +2,9 @@ var androidUtil = require('./androidUtil')
 var fs = require('fs')
 var iosUtil = require('./iosUtil')
 
-if (!fs.existsSync('./branch.json')) {
+if (!fs.existsSync('./branch.json') && !fs.existsSync('./branch.debug.json')) {
   return
 }
 
-androidUtil.addBranchJsonToAndroidAssetsFolder()
+androidUtil.addBranchConfigToAndroidAssetsFolder()
 iosUtil.addBranchConfigToXcodeProject()
